refactor: labelResolver.ts → labels.tsにリネーム
中身は辞書ルックアップなので「Resolver」は大げさ。
より実態に合った名前に変更。
